Update translate tests
Update translate test expected datasets to include the new
data.payload.reactions set on message contracts.

Change-type: patch
