+++ test/integration/webhooks/inbound-archive-message/expected.json
@@ -43,7 +43,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hello</div>\n"
+          "message": "<div>Hello</div>\n",
+          "reactions": {}
         }
       }
     },
@@ -175,7 +176,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Response!</div><br><div class=\"front-signature fa-uid_b7b4bdb0b19facd8150d852feeb4dda8\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div><br><blockquote type=\"cite\" class=\"front-blockquote\">On Wed, Oct 10, 2018 at 11:22 pm,  &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; Juan Cruz Viotti wrote:<br /><br /><div id=\"fae_1y9msqb-txqh3s\"><div>Hello</div>\n</div></blockquote>"
+          "message": "<div>Response!</div><br><div class=\"front-signature fa-uid_b7b4bdb0b19facd8150d852feeb4dda8\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div><br><blockquote type=\"cite\" class=\"front-blockquote\">On Wed, Oct 10, 2018 at 11:22 pm,  &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; Juan Cruz Viotti wrote:<br /><br /><div id=\"fae_1y9msqb-txqh3s\"><div>Hello</div>\n</div></blockquote>",
+          "reactions": {}
         }
       }
     },
