+++ test/integration/webhooks/inbound-archive/expected.json
@@ -43,7 +43,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hello</div>\n"
+          "message": "<div>Hello</div>\n",
+          "reactions": {}
         }
       }
     },
