+++ test/integration/webhooks/inbound-attachment/expected.json
@@ -80,7 +80,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Attachment test</p>\n</div>"
+          "message": "<div><p>Attachment test</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -106,7 +107,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "This is mirrored in [flowdock premium thread i2wIK9GpIvbofyoxnxENgGO1Jas](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/i2wIK9GpIvbofyoxnxENgGO1Jas)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=i2wIK9GpIvbofyoxnxENgGO1Jas&hmac=9ca94355a429a9ac18672da643178c325bb27f6a008cd9229455ff7ea9c4b7c0)"
+          "message": "This is mirrored in [flowdock premium thread i2wIK9GpIvbofyoxnxENgGO1Jas](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/i2wIK9GpIvbofyoxnxENgGO1Jas)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=i2wIK9GpIvbofyoxnxENgGO1Jas&hmac=9ca94355a429a9ac18672da643178c325bb27f6a008cd9229455ff7ea9c4b7c0)",
+          "reactions": {}
         }
       }
     },
@@ -164,7 +166,8 @@
               "bytesize": 1700
             }
           ],
-          "message": ""
+          "message": "",
+          "reactions": {}
         }
       }
     },
@@ -222,7 +225,8 @@
               "bytesize": 2818
             }
           ],
-          "message": ""
+          "message": "",
+          "reactions": {}
         }
       }
     },
