+++ test/integration/webhooks/inbound-comment-comment/expected.json
@@ -103,7 +103,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "First comment"
+          "message": "First comment",
+          "reactions": {}
         }
       }
     },
@@ -129,7 +130,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "Second comment"
+          "message": "Second comment",
+          "reactions": {}
         }
       }
     },
