+++ test/integration/webhooks/inbound-comment-edit/expected.json
@@ -80,7 +80,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Foo Bar</div>\n"
+          "message": "<div>Foo Bar</div>\n",
+          "reactions": {}
         }
       }
     },
@@ -106,7 +107,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "@test foo"
+          "message": "@test foo",
+          "reactions": {}
         }
       }
     },
