+++ test/integration/webhooks/inbound-delay-message/expected.json
@@ -43,7 +43,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hello</div>\n"
+          "message": "<div>Hello</div>\n",
+          "reactions": {}
         }
       }
     },
@@ -129,7 +130,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hey there!</div><br><div class=\"front-signature fa-uid_08dbba14523ca30f8679c237594b5e21\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div><br><blockquote type=\"cite\" class=\"front-blockquote\">On Wed, Oct 10, 2018 at 10:39 pm,  &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; Juan Cruz Viotti wrote:<br /><br /><div id=\"fae_1y9fb8z-ftkyv6\"><div>Hello</div>\n</div></blockquote>"
+          "message": "<div>Hey there!</div><br><div class=\"front-signature fa-uid_08dbba14523ca30f8679c237594b5e21\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div><br><blockquote type=\"cite\" class=\"front-blockquote\">On Wed, Oct 10, 2018 at 10:39 pm,  &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; Juan Cruz Viotti wrote:<br /><br /><div id=\"fae_1y9fb8z-ftkyv6\"><div>Hello</div>\n</div></blockquote>",
+          "reactions": {}
         }
       }
     },
