+++ test/integration/webhooks/inbound-delete-message/expected.json
@@ -43,7 +43,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hello</div>\n"
+          "message": "<div>Hello</div>\n",
+          "reactions": {}
         }
       }
     },
@@ -175,7 +176,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Response!</div><br><div class=\"front-signature fa-uid_1adb5fca3c6ea641296a83c893eddfa6\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div><br><blockquote type=\"cite\" class=\"front-blockquote\">On Wed, Oct 10, 2018 at 11:28 pm,  &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; Juan Cruz Viotti wrote:<br /><br /><div id=\"fae_1y9nrkr-xcwv15\"><div>Hello</div>\n</div></blockquote>"
+          "message": "<div>Response!</div><br><div class=\"front-signature fa-uid_1adb5fca3c6ea641296a83c893eddfa6\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div><br><blockquote type=\"cite\" class=\"front-blockquote\">On Wed, Oct 10, 2018 at 11:28 pm,  &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; Juan Cruz Viotti wrote:<br /><br /><div id=\"fae_1y9nrkr-xcwv15\"><div>Hello</div>\n</div></blockquote>",
+          "reactions": {}
         }
       }
     },
