+++ test/integration/webhooks/inbound-inbound-inbound/expected.json
@@ -80,8 +80,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>One</div><br /><div class=\"gmail_quote\"><div>On Mon, Oct 8, 2018 at 7:39 PM Juan Cruz Viotti &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; wrote:<br /></div><blockquote class=\"gmail_quote front-blockquote\" style=\"margin:0 0 0 .8ex;border-left:1px #ccc solid;padding-left:1ex\"><div><br /></div>\n</blockquote></div>\n"
-
+          "message": "<div>One</div><br /><div class=\"gmail_quote\"><div>On Mon, Oct 8, 2018 at 7:39 PM Juan Cruz Viotti &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; wrote:<br /></div><blockquote class=\"gmail_quote front-blockquote\" style=\"margin:0 0 0 .8ex;border-left:1px #ccc solid;padding-left:1ex\"><div><br /></div>\n</blockquote></div>\n",
+          "reactions": {}
         }
       }
     },
@@ -133,8 +133,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Two</div><br /><div class=\"gmail_quote\"><div>On Mon, Oct 8, 2018 at 7:39 PM Juan Cruz Viotti &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; wrote:<br /></div><blockquote class=\"gmail_quote front-blockquote\" style=\"margin:0 0 0 .8ex;border-left:1px #ccc solid;padding-left:1ex\"><div>One</div><br /><div class=\"gmail_quote\"><div>On Mon, Oct 8, 2018 at 7:39 PM Juan Cruz Viotti &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; wrote:<br /></div><blockquote class=\"gmail_quote front-blockquote\" style=\"margin:0 0 0 .8ex;border-left:1px #ccc solid;padding-left:1ex\"><div><br /></div>\n</blockquote></div></blockquote></div>\n"
-
+          "message": "<div>Two</div><br /><div class=\"gmail_quote\"><div>On Mon, Oct 8, 2018 at 7:39 PM Juan Cruz Viotti &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; wrote:<br /></div><blockquote class=\"gmail_quote front-blockquote\" style=\"margin:0 0 0 .8ex;border-left:1px #ccc solid;padding-left:1ex\"><div>One</div><br /><div class=\"gmail_quote\"><div>On Mon, Oct 8, 2018 at 7:39 PM Juan Cruz Viotti &lt;<a href=\"mailto:juan@resin.io\" target=\"_blank\" rel=\"noopener noreferrer\">juan@resin.io</a>&gt; wrote:<br /></div><blockquote class=\"gmail_quote front-blockquote\" style=\"margin:0 0 0 .8ex;border-left:1px #ccc solid;padding-left:1ex\"><div><br /></div>\n</blockquote></div></blockquote></div>\n",
+          "reactions": {}
         }
       }
     },
