+++ test/integration/webhooks/inbound-tag-tag/expected.json
@@ -46,8 +46,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hello</div>\n"
-
+          "message": "<div>Hello</div>\n",
+          "reactions": {}
         }
       }
     },
