+++ test/integration/webhooks/intercom-draft-message/expected.json
@@ -43,7 +43,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>JF test (ignore)</p>\n</div>"
+          "message": "<div><p>JF test (ignore)</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -106,7 +107,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "This is mirrored in [flowdock premium thread kUAq5jWTJ6_lXpXcb3dT0gDO8Gc](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/kUAq5jWTJ6_lXpXcb3dT0gDO8Gc)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=kUAq5jWTJ6_lXpXcb3dT0gDO8Gc&hmac=c57c67c40a1aa89afaa82569c5f9ecfa6e174a17f494dc49424253c371b36424)"
+          "message": "This is mirrored in [flowdock premium thread kUAq5jWTJ6_lXpXcb3dT0gDO8Gc](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/kUAq5jWTJ6_lXpXcb3dT0gDO8Gc)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=kUAq5jWTJ6_lXpXcb3dT0gDO8Gc&hmac=c57c67c40a1aa89afaa82569c5f9ecfa6e174a17f494dc49424253c371b36424)",
+          "reactions": {}
         }
       }
     },
