+++ test/integration/webhooks/intercom-inbound-archive-email-prefix/expected.json
@@ -43,7 +43,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Test test 2</p>\n</div>"
+          "message": "<div><p>Test test 2</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -106,7 +107,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "This is mirrored in [flowdock premium thread 3o8hhRVDQOwJwh-u9_2Sq7cMYi4](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/3o8hhRVDQOwJwh-u9_2Sq7cMYi4)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=3o8hhRVDQOwJwh-u9_2Sq7cMYi4&hmac=18535b24330510d884ec4bd6c4327d2cc20a9971bbf12e9a6c96c4ee994a42dc)"
+          "message": "This is mirrored in [flowdock premium thread 3o8hhRVDQOwJwh-u9_2Sq7cMYi4](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/3o8hhRVDQOwJwh-u9_2Sq7cMYi4)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=3o8hhRVDQOwJwh-u9_2Sq7cMYi4&hmac=18535b24330510d884ec4bd6c4327d2cc20a9971bbf12e9a6c96c4ee994a42dc)",
+          "reactions": {}
         }
       }
     },
