+++ test/integration/webhooks/intercom-no-author/expected.json
@@ -81,8 +81,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<p>Hi dt-rush,<br />To run the commands on the host OS what we do is generate bash scripts at runtime using the users parameters (stored in /data) and before we reboot the device we run the script.  So they are running from the shell using the Command::new() portion of std::process::Command from Rust.  </p>\n<p>For some additional information about how the dockerfile looks here are the packages we are pulling in as well as the NetworkManager.service mask:</p>\n<p>---Snip---<br />FROM balenalib/amd64-ubuntu</p>\n<p>ENV INITSYSTEM on</p>\n<p>RUN apt-get update</p>\n<p>RUN apt-get install -y dnsmasq wireless-tools vim file udhcpd network-manager \\<br />    &amp;&amp; systemctl mask NetworkManager.service \\<br />    &amp;&amp; apt-get clean \\<br />    &amp;&amp; rm -rf /var/lib/apt/lists/*<br />---Snip---</p>\n<p>The rest is just some application specific stuff to put files in the container and run it.</p>\n<p>Ill look into the systemd service idea today.  One issue I am assuming I&#39;d need to solve is bringing up the new wifi interface before I reboot, I&#39;m not sure if NM will allow me to do that if the balena tunnel is running on another device however.  If you have any other suggestions or details around that approach I&#39;m listening.</p>\n<p>Thanks,<br />Brant</p>\n<a href=\"https://www.balena-cloud.com?hidden=reply&source=flowdock&flow=rulemotion/public-s-community&thread=TAz9aNXmYye6c4FeZGg-7MmHznc&hmac=d430d2534f80fa099bd2556deb1facc82d5dff927a1d6bec08e33a33e81466fc\" target=\"_blank\" rel=\"noopener noreferrer\"></a>"
-
+          "message": "<p>Hi dt-rush,<br />To run the commands on the host OS what we do is generate bash scripts at runtime using the users parameters (stored in /data) and before we reboot the device we run the script.  So they are running from the shell using the Command::new() portion of std::process::Command from Rust.  </p>\n<p>For some additional information about how the dockerfile looks here are the packages we are pulling in as well as the NetworkManager.service mask:</p>\n<p>---Snip---<br />FROM balenalib/amd64-ubuntu</p>\n<p>ENV INITSYSTEM on</p>\n<p>RUN apt-get update</p>\n<p>RUN apt-get install -y dnsmasq wireless-tools vim file udhcpd network-manager \\<br />    &amp;&amp; systemctl mask NetworkManager.service \\<br />    &amp;&amp; apt-get clean \\<br />    &amp;&amp; rm -rf /var/lib/apt/lists/*<br />---Snip---</p>\n<p>The rest is just some application specific stuff to put files in the container and run it.</p>\n<p>Ill look into the systemd service idea today.  One issue I am assuming I&#39;d need to solve is bringing up the new wifi interface before I reboot, I&#39;m not sure if NM will allow me to do that if the balena tunnel is running on another device however.  If you have any other suggestions or details around that approach I&#39;m listening.</p>\n<p>Thanks,<br />Brant</p>\n<a href=\"https://www.balena-cloud.com?hidden=reply&source=flowdock&flow=rulemotion/public-s-community&thread=TAz9aNXmYye6c4FeZGg-7MmHznc&hmac=d430d2534f80fa099bd2556deb1facc82d5dff927a1d6bec08e33a33e81466fc\" target=\"_blank\" rel=\"noopener noreferrer\"></a>",
+          "reactions": {}
         }
       }
     }
