+++ test/integration/webhooks/intercom-recap/expected.json
@@ -80,7 +80,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Hi, is there a quick and easy way to revert to an older release? I've done it succesfully by &quot;pushing&quot; the older commit but it doesn't seem to be working for me right now</p>\n</div>"
+          "message": "<div><p>Hi, is there a quick and easy way to revert to an older release? I've done it succesfully by &quot;pushing&quot; the older commit but it doesn't seem to be working for me right now</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -106,7 +107,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "This is mirrored in [flowdock premium thread m-PT2l8ed9u-l_3FKX3kNhXzq5g](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/m-PT2l8ed9u-l_3FKX3kNhXzq5g)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=m-PT2l8ed9u-l_3FKX3kNhXzq5g&hmac=4938cfcb9c79d694fdfb1fc8cd4eb29f61dfc49959f85590f8c9327b69b2ca95)"
+          "message": "This is mirrored in [flowdock premium thread m-PT2l8ed9u-l_3FKX3kNhXzq5g](https://www.flowdock.com/app/rulemotion/public-s-premium/threads/m-PT2l8ed9u-l_3FKX3kNhXzq5g)\n[](https://www.balena-cloud.com?hidden=murmur&source=flowdock&flow=rulemotion/public-s-premium&thread=m-PT2l8ed9u-l_3FKX3kNhXzq5g&hmac=4938cfcb9c79d694fdfb1fc8cd4eb29f61dfc49959f85590f8c9327b69b2ca95)",
+          "reactions": {}
         }
       }
     },
@@ -132,7 +134,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Never mind I think I've gotten that to work again.. lol</p>\n</div>"
+          "message": "<div><p>Never mind I think I've gotten that to work again.. lol</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -158,7 +161,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Is this something I can do on the balena dashboard for future reference?</p>\n</div>"
+          "message": "<div><p>Is this something I can do on the balena dashboard for future reference?</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -184,7 +188,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Hey, it's not actually implemented on the dashboard side, but you can use staged releases, with the code found in examples here: <a href=\"https://github.com/balena-io-projects/staged-releases\">https://github.com/balena-io-projects/staged-releases</a></p>\n</div>"
+          "message": "<div><p>Hey, it's not actually implemented on the dashboard side, but you can use staged releases, with the code found in examples here: <a href=\"https://github.com/balena-io-projects/staged-releases\">https://github.com/balena-io-projects/staged-releases</a></p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -210,7 +215,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>I meant to say it's not implemented <em>yet</em>, but it is coming.</p>\n</div>"
+          "message": "<div><p>I meant to say it's not implemented <em>yet</em>, but it is coming.</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -236,7 +242,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "#summary user wanted to know how to rollback a device, showed them staged-releases"
+          "message": "#summary user wanted to know how to rollback a device, showed them staged-releases",
+          "reactions": {}
         }
       }
     },
@@ -262,7 +269,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>Sounds good, thanks!</p>\n</div>"
+          "message": "<div><p>Sounds good, thanks!</p>\n</div>",
+          "reactions": {}
         }
       }
     },
@@ -288,7 +296,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "#summary user wanted to know how to rollback a device, showed them staged-releases (prev status)"
+          "message": "#summary user wanted to know how to rollback a device, showed them staged-releases (prev status)",
+          "reactions": {}
         }
       }
     },
@@ -314,7 +323,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "closed"
+          "message": "closed",
+          "reactions": {}
         }
       }
     },
@@ -340,7 +350,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "Test"
+          "message": "Test",
+          "reactions": {}
         }
       }
     },
