+++ test/integration/webhooks/outbound/expected.json
@@ -80,7 +80,8 @@
           "alertsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div>Hello</div><br><div class=\"front-signature fa-uid_3261ad35206a43895d965cefeb9a22c2\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div>"
+          "message": "<div>Hello</div><br><div class=\"front-signature fa-uid_3261ad35206a43895d965cefeb9a22c2\">—<br />Juan Cruz Viotti,<div>Software Engineer, Resin.io (https://resin.io)</div><div>twt: @resin_io</div></div>",
+          "reactions": {}
         }
       }
     }
