+++ test/integration/webhooks/reminder/expected.json
@@ -79,7 +79,8 @@
           "mentionsUser": [],
           "mentionsGroup": [],
           "alertsGroup": [],
-          "message": "<div><p>No worries. I've worked around it for now by pinning to an older base image.</p>\n</div>"
+          "message": "<div><p>No worries. I've worked around it for now by pinning to an older base image.</p>\n</div>",
+          "reactions": {}
         },
         "timestamp": "2018-12-18T18:32:53.000Z"
       }
